+++ API/Controllers/TaskController.cs
@@ -1,5 +1,4 @@
-﻿using Core.Entities;
-using Core.Entities.DTO;
+﻿using Core.Entities.DTO;
 using Core.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -20,70 +19,87 @@ namespace API.Controllers
         [HttpPost("CreateTask")]
         public async Task<IActionResult> CreateTask([FromBody] TaskDto taskDto)
         {
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
-
+            var userId = GetUserIdByClaimTypes();
+            if (userId == null)
+                return Unauthorized();
             await _taskService.CreateTask(taskDto, userId);
             return Ok();
         }
 
         [Authorize]
         [HttpGet("Tasks")]
-        public async Task<List<TaskEntity>> GetTasks()
+        public async Task<IActionResult> GetTasks()
         {
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var userId = GetUserIdByClaimTypes();
+            if (userId == null)
+                return Unauthorized();
+            var tasks = await _taskService.GetTasks(userId);
+            if (!tasks.Any())
+                return NotFound("No tasks found for the user.");
 
-            var task = await _taskService.GetTasks(userId);
-            return task;
+            return Ok(tasks);
         }
 
         [Authorize]
         [HttpGet("Tasks/done")]
-        public async Task<List<TaskEntity>> GetDoneTasks()
+        public async Task<IActionResult> GetDoneTasks()
         {
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
-
+            var userId = GetUserIdByClaimTypes();
+            if (userId == null)
+                return Unauthorized();
             var task = await _taskService.GetDoneTasks(userId);
-            return task;
+            return Ok(task);
         }
 
         [Authorize]
         [HttpGet("Tasks/notdone")]
-        public async Task<List<TaskEntity>> GetNotDoneTasks()
+        public async Task<IActionResult> GetNotDoneTasks()
         {
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
-
+            var userId = GetUserIdByClaimTypes();
+            if (userId == null)
+                return Unauthorized();
             var task = await _taskService.GetNotDoneTasks(userId);
-            return task;
+            return Ok(task);
         }
 
         [Authorize]
         [HttpPut("EditTask/{id}")]
-        public async Task<TaskEntity> EditTask(int id, [FromBody] TaskDto taskDto)
+        public async Task<IActionResult> EditTask(int id, [FromBody] TaskDto taskDto)
         {
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
-
+            var userId = GetUserIdByClaimTypes();
+            if (userId == null)
+                return Unauthorized();
             var task = await _taskService.EditTask(userId, id, taskDto);
-            return task;
+            return Ok(task);
         }
 
         [Authorize]
         [HttpPut("StatusTask/{id}")]
-        public async Task<bool> EditTask(int id, [FromBody] StatusDto status)
+        public async Task<IActionResult> EditTask(int id, [FromBody] StatusDto status)
         {
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var userId = GetUserIdByClaimTypes();
+            if (userId == null)
+                return Unauthorized();
 
             var task = await _taskService.ChangeTaskStatus(userId, id, status);
-            return task;
+            return Ok(task);
         }
 
         [Authorize]
         [HttpDelete("RemoveTask/{id}")]
-        public async Task<bool> RemoveTask(int id)
+        public async Task<IActionResult> RemoveTask(int id)
+        {
+            var userId = GetUserIdByClaimTypes();
+            if (userId == null)
+                return Unauthorized();
+            var task = await _taskService.RemoveTask(userId, id);
+            return Ok(task);
+        }
+
+        private string? GetUserIdByClaimTypes()
         {
             var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
-
-            var task = await _taskService.RemoveTask(userId, id);
-            return task;
+            return userId;
         }
     }
 }
